Metida alerta en eliminar menú del dashboard

+++ src/app/modules/Dashboard/components/db-menu-list/db-menu-list.component.ts
@@ -2,6 +2,7 @@ import { Component, inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { Menu } from 'src/app/core/interfaces/menu.interface';
 import { MenuService } from 'src/app/core/services/menu.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-db-menu-list',
@@ -21,32 +22,58 @@ export class DbMenuListComponent {
 
   async ngOnInit() {
     try {
-      const response = await this.menuService.getAll()      
+      const response = await this.menuService.getAll()
       this.menus = response
-      
+
     } catch (error) {
       console.log(error);
-      
+
     }
   }
 
   async erase(menu: Menu) {
-    try {
-      const response = await this.menuService.delete(Number(menu.id))
-      
-      const response2 = await this.menuService.getAll()      
-      this.menus = response2
-      
-    } catch (error) {
-      console.log(error);      
-    }
+
+    Swal.fire({
+      title: "¿Quieres eliminar este menú?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+      confirmButtonColor: "var(--secondary-color)",
+      cancelButtonColor: "var(--third-color)",
+      reverseButtons: true,
+      color: "var(--main-color)",
+      background: "var(--bg-color)"
+
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const response = await this.menuService.delete(Number(menu.id))
+
+          const response2 = await this.menuService.getAll()
+          this.menus = response2
+
+        } catch (error) {
+          console.log(error);
+        }
+
+        Swal.fire({
+          title: "Reserva cancelada",
+          icon: "success",
+          confirmButtonColor: "var(--secondary-color)",
+          color: "var(--main-color)",
+          background: "var(--bg-color)"
+        });
+      }
+    });
+
   }
 
   async edit(menu: Menu) {
-    
+
     const id = menu.id?.toString()
-    this.router.navigate(["/dashboard/menuedit",id])
-    
+    this.router.navigate(["/dashboard/menuedit", id])
+
   }
